requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
mcp>=1.0.0
aiohttp>=3.9.0
streamlit>=1.28.0
streamlit-chat>=0.1.1

# FastAPI and web server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.1
aiosqlite>=0.19.0
itsdangerous>=2.1.2

# Data pipeline dependencies
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
python-multipart==0.0.6
